+++ CHANGELOG.md
@@ -5,13 +5,15 @@
 * **[Breaking]**: `inc` folder renamed to `src` as there are may be additional extra includes
 
 * **[Feat]**: added `copyrights()` filter
+* **[Feat]**: added `Mail` facade to send emails via SMTP or default WordPress transport
+* **[Feat]**: added default mail layout
 * **[Feat]**: added `sprite()` function for SVG icons like `<svg><use xlink:href="{{ sprite('sprite-id') }}" /></svg>`
 * **[Refactor]**: `Mod` facade now implements simple settings creation with `Mod::type('id', 'label')`
 * **[Build]**: added `.min` postfix for minified scripts and styles
 * **[Build]**: added `.stylelintignore` and `.eslintignore` files
 * **[Build]**: changed `.browserslistrc` config
 
-Release Date: September 10, 2021
+Release Date: September 12, 2021
 
 ## Release 0.14.2